requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlmodel==0.0.14

# Spotify API (for search)
spotipy==2.23.0

# HTTP requests (for auth flow)
requests>=2.31.0

# QR Code generation
qrcode[pil]==7.4.2
Pillow>=10.1.0

# Environment variables
python-dotenv==1.0.0

# WebSocket support (included in FastAPI/Starlette)
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
